add support for custom cover images
In particular, if custom cover images are enabled via
CustomCoverImagesURL, they can be displayed by CoverFlow. If a bib
has a local cover image, that will take precedence over the
custom cover.

Note that the OPACCustomCoverImages system preference is ignored for
now; as long as a custom cover image URL is available, it
will be considered.

Signed-off-by: Galen Charlton <gmc@equinoxOLI.org>